backend/requirements.txt
asgiref==3.6.0
Django==4.2
django-cors-headers==3.14.0
django-environ==0.10.0
djangorestframework==3.14.0
mysqlclient==2.1.1
pytz==2023.3
six==1.16.0
sqlparse==0.4.3
